@@ -1,22 +1,15 @@
 using Microsoft.AspNetCore.Mvc;
-using Graphitie.Models; 
 using Graphitie.Services;
 
 namespace Graphitie.Controllers.Microsoft;
 
 [ApiController]
 [Route("[controller]")]
-public class ContactsController : ControllerBase
+public class ContactsController(ILogger<ContactsController> logger, GraphitieService graphitieService) : ControllerBase
 {
-    private readonly ILogger<ContactsController> _logger;
+    private readonly ILogger<ContactsController> _logger = logger;
 
-    private readonly GraphitieService _graphitieService;
-
-    public ContactsController(ILogger<ContactsController> logger, GraphitieService graphitieService)
-    {
-        _logger = logger;
-        _graphitieService = graphitieService;
-    }
+    private readonly GraphitieService _graphitieService = graphitieService;
 
     [HttpGet(Name = "CopyMemberContacts")]
     public async Task CopyMemberContacts(string userId, string folderName, string? birthdaySiteId = null)
@@ -25,8 +18,8 @@ public class ContactsController : ControllerBase
     }
 
     [HttpPost(Name = "AddContacts")]
-    public async Task AddContacts(string userId, string folderName, [FromBody] List<Contact> contacts)
+    public void AddContacts(string userId, string folderName)
     {
-        await _graphitieService.AddContacts(userId, folderName);
+        _graphitieService.AddContacts(userId, folderName);
     }
 }
