@@ -6,17 +6,11 @@ namespace Graphitie.Controllers.Microsoft;
 
 [ApiController]
 [Route("[controller]")]
-public class EmailController : ControllerBase
+public class EmailController(ILogger<EmailController> logger, GraphitieService graphitieService) : ControllerBase
 {
-    private readonly ILogger<EmailController> _logger;
+    private readonly ILogger<EmailController> _logger = logger;
 
-    private readonly GraphitieService _graphitieService;
-
-    public EmailController(ILogger<EmailController> logger, GraphitieService graphitieService)
-    {
-        _logger = logger;
-        _graphitieService = graphitieService;
-    }
+    private readonly GraphitieService _graphitieService = graphitieService;
 
     [HttpPut(Name = "Email")]  
     public async Task SendMail(string user, [FromBody] Mail email)
